Refactor: OptionOverlay 파일추가 후 선택옵션 코드 분리
